refactor(charts): share viewport primitives
