+++ Radix/Features/Visualization/ChartViewportInteractionView.swift
@@ -1,0 +1,225 @@
+import AppKit
+
+@MainActor
+class ChartViewportInteractionView: ChartKeyboardInteractionView, NSDraggingSource {
+    var onHover: (CGPoint?) -> Void = { _ in }
+    var onClick: (CGPoint, Int) -> Void = { _, _ in }
+    var onPan: (CGSize) -> Void = { _ in }
+    var onMagnify: (CGPoint, CGFloat) -> Void = { _, _ in }
+    var canStartPan: (CGPoint) -> Bool = { _ in false }
+    var onDragActiveChange: (Bool) -> Void = { _ in }
+    var help: (CGPoint) -> String? = { _ in nil }
+    var isPanEnabled = false
+
+    private static let dragThreshold: CGFloat = 3
+    private static let lineScrollScale: CGFloat = 10
+    private static let maximumScrollPanDelta: CGFloat = 80
+
+    private var trackingArea: NSTrackingArea?
+    private var mouseDownLocation: CGPoint?
+    private var lastDragLocation: CGPoint?
+    private var shouldPanFromMouseDownLocation = false
+    private var didPan = false
+    private var didStartDrag = false
+
+    override var isFlipped: Bool {
+        true
+    }
+
+    override func updateTrackingAreas() {
+        super.updateTrackingAreas()
+
+        if let trackingArea {
+            removeTrackingArea(trackingArea)
+        }
+
+        let trackingArea = NSTrackingArea(
+            rect: .zero,
+            options: [.activeInKeyWindow, .inVisibleRect, .mouseEnteredAndExited, .mouseMoved],
+            owner: self,
+            userInfo: nil
+        )
+        addTrackingArea(trackingArea)
+        self.trackingArea = trackingArea
+    }
+
+    override func mouseEntered(with event: NSEvent) {
+        updatePointerFeedback(at: eventLocation(event))
+    }
+
+    override func mouseMoved(with event: NSEvent) {
+        updatePointerFeedback(at: eventLocation(event))
+    }
+
+    override func mouseExited(with event: NSEvent) {
+        onHover(nil)
+        toolTip = nil
+    }
+
+    override func mouseDown(with event: NSEvent) {
+        focusForKeyboardInput()
+        let location = eventLocation(event)
+        mouseDownLocation = location
+        lastDragLocation = location
+        shouldPanFromMouseDownLocation = isPanEnabled && canStartPan(location)
+        didPan = false
+        didStartDrag = false
+    }
+
+    override func mouseDragged(with event: NSEvent) {
+        guard let mouseDownLocation,
+              let lastDragLocation else { return }
+
+        let location = eventLocation(event)
+        if !didPan {
+            guard didExceedDragThreshold(from: mouseDownLocation, to: location) else {
+                return
+            }
+            didPan = true
+        }
+
+        if shouldPanFromMouseDownLocation {
+            defer { self.lastDragLocation = location }
+            guard isPanEnabled else { return }
+
+            onPan(CGSize(
+                width: location.x - lastDragLocation.x,
+                height: location.y - lastDragLocation.y
+            ))
+            updatePointerFeedback(at: location)
+            return
+        }
+
+        if !didStartDrag,
+           let draggingItem = draggingItem(at: mouseDownLocation) {
+            didStartDrag = true
+            onDragActiveChange(true)
+            beginDraggingSession(with: [draggingItem], event: event, source: self)
+            return
+        }
+
+        defer { self.lastDragLocation = location }
+        guard isPanEnabled else { return }
+
+        onPan(CGSize(
+            width: location.x - lastDragLocation.x,
+            height: location.y - lastDragLocation.y
+        ))
+        updatePointerFeedback(at: location)
+    }
+
+    override func mouseUp(with event: NSEvent) {
+        let location = eventLocation(event)
+        if !didPan {
+            onClick(location, event.clickCount)
+        }
+        mouseDownLocation = nil
+        lastDragLocation = nil
+        shouldPanFromMouseDownLocation = false
+        didPan = false
+        didStartDrag = false
+    }
+
+    override func magnify(with event: NSEvent) {
+        let location = eventLocation(event)
+        onMagnify(location, max(0.75, 1 + event.magnification))
+        updatePointerFeedback(at: location)
+    }
+
+    override func scrollWheel(with event: NSEvent) {
+        let location = eventLocation(event)
+        let zoomModifiers: NSEvent.ModifierFlags = [.command, .option]
+
+        if !event.modifierFlags.intersection(zoomModifiers).isEmpty {
+            let scrollDelta = event.scrollingDeltaY != 0
+                ? event.scrollingDeltaY
+                : -event.scrollingDeltaX
+            guard scrollDelta != 0 else { return }
+
+            onMagnify(location, pow(1.0025, scrollDelta))
+            updatePointerFeedback(at: location)
+            return
+        }
+
+        if isPanEnabled {
+            guard let panDelta = panDelta(for: event) else { return }
+            onPan(panDelta)
+            updatePointerFeedback(at: location)
+            return
+        }
+
+        super.scrollWheel(with: event)
+    }
+
+    func draggingItem(at location: CGPoint) -> NSDraggingItem? {
+        nil
+    }
+
+    func draggingSession(
+        _ session: NSDraggingSession,
+        sourceOperationMaskFor context: NSDraggingContext
+    ) -> NSDragOperation {
+        .copy
+    }
+
+    func ignoreModifierKeys(for session: NSDraggingSession) -> Bool {
+        true
+    }
+
+    func draggingSession(
+        _ session: NSDraggingSession,
+        endedAt screenPoint: NSPoint,
+        operation: NSDragOperation
+    ) {
+        onDragActiveChange(false)
+    }
+
+    private func updatePointerFeedback(at location: CGPoint) {
+        onHover(location)
+        toolTip = help(location)
+    }
+
+    private func eventLocation(_ event: NSEvent) -> CGPoint {
+        convert(event.locationInWindow, from: nil)
+    }
+
+    private func didExceedDragThreshold(from start: CGPoint, to end: CGPoint) -> Bool {
+        let dx = end.x - start.x
+        let dy = end.y - start.y
+        return ((dx * dx) + (dy * dy)) >= (Self.dragThreshold * Self.dragThreshold)
+    }
+
+    private func panDelta(for event: NSEvent) -> CGSize? {
+        var delta = CGSize(
+            width: event.scrollingDeltaX,
+            height: event.scrollingDeltaY
+        )
+
+        guard delta != .zero else { return nil }
+
+        if !event.isDirectionInvertedFromDevice {
+            delta.width *= -1
+            delta.height *= -1
+        }
+
+        if !event.hasPreciseScrollingDeltas {
+            delta.width *= Self.lineScrollScale
+            delta.height *= Self.lineScrollScale
+        }
+
+        return CGSize(
+            width: delta.width.clamped(
+                to: -Self.maximumScrollPanDelta...Self.maximumScrollPanDelta
+            ),
+            height: delta.height.clamped(
+                to: -Self.maximumScrollPanDelta...Self.maximumScrollPanDelta
+            )
+        )
+    }
+}
+
+private extension Comparable {
+    func clamped(to range: ClosedRange<Self>) -> Self {
+        Swift.min(Swift.max(self, range.lowerBound), range.upperBound)
+    }
+}
